+++ src/index.ts
@@ -35,68 +35,11 @@ require('./routes/submitMatchData')(app, dbHandler, auth);
 require('./routes/checkUser')(app, dbHandler, auth);
 require('./routes/fetchScouterSuggestions')(app, dbHandler);
 require('./routes/fetchScouterUIDs')(app, dbHandler);
+require('./routes/fetchPitConfig')(app);
+require('./routes/findTeamNickname')(app, dbHandler);
+require('./routes/fetchAllTeamNicknamesAtCompetition')(app, dbHandler);
+require('./routes/fetchMatchConfig')(app);
 
-
-/**
- * GET route '/api/fetchAllTeamNicknamesAtCompetition'
- * Allows the application to fetch the nicknames for all the teams which are at a competition. (For example, Team 2022 = Titan Robotics)
- * @param competition is the identifier for the competition: e.g. '2020ilch'.
- * @returns back to the client let resobj (competition id, JSON of the team number and nicknames) and HTTP Status Code 200 OK.
- */
-app.get('/api/fetchAllTeamNicknamesAtCompetition', async (req: any, res:any) => {
-  let val;
-  const competition = String(req.query.competition);
-  try {
-    val = await dbHandler.fetchAllTeamNicknamesAtCompetition(req.db, competition).catch((e) => { console.error(e); val.err_occur = true; });
-  } catch (e) {
-    console.error(e);
-    val.err_occur = true;
-  }
-  let resobj = null;
-  if (val.err_occur === false) {
-    resobj = {
-      success: true,
-      competition,
-      data: val.data,
-    };
-  } else {
-    resobj = {
-      success: false,
-      reasons: val.err_reasons,
-    };
-  }
-  res.json(resobj);
-});
-/**
- * GET route '/api/findTeamNickname'
- * Allows the application to get the nickname for a team, given the team number.
- * @param teamNum is the FRC team number: e.g. '2022'.
- * @returns back to the client let resobj (team number and nickname) and HTTP Status Code 200 OK.
- */
-app.get('/api/findTeamNickname', async (req: any, res:any) => {
-  let val;
-  const teamNumber = req.query.team_number;
-  try {
-    val = await dbHandler.findTeamNickname(req.db, teamNumber).catch((e) => { console.error(e); val.err_occur = true; });
-  } catch (e) {
-    console.error(e);
-    val.err_occur = true;
-  }
-  let resobj = null;
-  if (val.err_occur === false) {
-    resobj = {
-      success: true,
-      teamNum: teamNumber,
-      nickname: val.data,
-    };
-  } else {
-    resobj = {
-      success: false,
-      reasons: val.err_reasons,
-    };
-  }
-  res.json(resobj);
-});
 /**
  * GET route '/api/fetchCompetitionSchedule'
  * Allows the application to get all the matches for a given competition.
@@ -435,185 +378,6 @@ app.get('/privacy-policy', (req: any, res:any) => {
   res.status(301).redirect('https://drive.google.com/a/imsa.edu/file/d/11_cAuaerCrQ3BBXNx_G_zw1ZyGaTWx0z/view?usp=sharing');
 });
 
-app.get('/api/fetchMatchConfig', async (req: any, res:any) => {
-  const response = [
-    {
-      Auto: [
-        {
-          name: 'Passed Auto Line?',
-          key: 'pass-line',
-          widget: 'segment',
-          options: ['Don\'t Know', 'No', 'Yes'],
-        },
-        {
-          name: 'Initial Balls Stored',
-          key: 'balls-started',
-          widget: 'stepper',
-        },
-        {
-          name: 'Extra Balls Collected',
-          key: 'balls-collected',
-          widget: 'stepper',
-        },
-        {
-          name: 'Balls Scored Upper',
-          key: 'balls-upper-auto',
-          widget: 'stepper',
-        },
-        {
-          name: 'Balls Scored Lower',
-          key: 'balls-lower-auto',
-          widget: 'stepper',
-        },
-      ],
-    },
-    {
-      Teleop: [
-        {
-          name: 'Spun Wheel?',
-          key: 'spun-wheel',
-          widget: 'segment',
-          options: ['Don\'t know', 'No', 'Position', 'Color'],
-        },
-        {
-          name: 'Balls Scored Upper',
-          key: 'balls-upper-teleop',
-          widget: 'stepper',
-        },
-        {
-          name: 'Balls Scored Lower',
-          key: 'balls-lower-teleop',
-          widget: 'stepper',
-        },
-        {
-          name: 'Did they shoot from a vulnerable location??',
-          key: 'shooting-vulnerable',
-          widget: 'segment',
-          options: ['Don\'t know', 'No', 'Yes'],
-        },
-        {
-          name: 'Where could they shoot from?',
-          key: 'shooting-notes',
-          widget: 'text-area',
-        },
-        {
-          name: 'Did they climb?',
-          key: 'climb',
-          widget: 'segment',
-          options: ['Don\'t know', 'No Attempt', 'Failed', 'Yes'],
-        },
-        {
-          name: 'Did they play defense?',
-          key: 'defense',
-          widget: 'segment',
-          options: ['Don\'t know', 'No', 'Yes'],
-        },
-        {
-          name: 'What teams did this one play defense on?',
-          key: 'defense-notes',
-          widget: 'text-area',
-        },
-      ],
-    },
-    {
-      Notes: [
-        {
-          name: 'Overall Competency',
-          key: 'competency',
-          widget: 'segment',
-          options: ['Don\'t know', 'Awful', 'Meh', 'Good', 'Best'],
-        },
-        {
-          name: 'Speed',
-          key: 'speed',
-          widget: 'segment',
-          options: ['Don\'t know', 'Slow', 'Med.', 'Fast', 'Ludicrous'],
-        },
-        {
-          name: 'Strategic Focus',
-          key: 'strategic-focus',
-          widget: 'segment',
-          options: ['Don\'t know', 'Offense', 'Defense', 'Hybrid'],
-        },
-        {
-          name: 'How could we use this robot in a strategy?',
-          key: 'strategy-notes',
-          widget: 'text-area',
-        },
-      ],
-    },
-  ];
-  res.json(response);
-});
-
-app.get('/api/fetchPitConfig', async (req: any, res:any) => {
-  const response = [
-    {
-      Pit: [
-        {
-          name: 'Updated on Match',
-          key: 'match-updated',
-          widget: 'stepper',
-        },
-        {
-          name: 'Can do low balls?',
-          key: 'low-balls',
-          widget: 'segment',
-          options: ['Don\'t Know', 'Yes', 'No'],
-        },
-        {
-          name: 'Can do high balls?',
-          key: 'high-balls',
-          widget: 'segment',
-          options: ['Don\'t Know', 'Yes', 'No'],
-        },
-        {
-          name: 'Has wheel mechanism?',
-          key: 'wheel-mechanism',
-          widget: 'segment',
-          options: ['Don\'t Know', 'Yes', 'No'],
-        },
-        {
-          name: 'Demonstrated wheel success?',
-          key: 'wheel-success',
-          widget: 'segment',
-          options: ['Don\'t Know', 'Yes', 'No'],
-        },
-        {
-          name: 'Strategic focus',
-          key: 'strategic-focus',
-          widget: 'segment',
-          options: ['idk', 'Offense', 'Defense', 'Hybrid'],
-        },
-        {
-          name: 'Climb mechanism',
-          key: 'climb-mechanism',
-          widget: 'segment',
-          options: ['Don\'t Know', 'x0', 'x1', 'x2', 'x3'],
-
-        },
-        {
-          name: 'Climb requirements (space? time?)',
-          key: 'climb-requirements',
-          widget: 'text-area',
-        },
-        {
-          name: 'Attitude toward Titan Robotics',
-          key: 'attitude',
-          widget: 'segment',
-          options: ['Don\'t Know', 'Negative', 'Neutral', 'Positive', 'Love'],
-
-        },
-        {
-          name: 'Other notes',
-          key: 'defense-notes',
-          widget: 'text-area',
-        },
-      ],
-    },
-  ];
-  res.json(response);
-});
 
 app.post('/api/submitPitData', auth.checkAuth, async (req: any, res:any) => {
   let val;
